Add files via upload
DESCRIPTION

Problem:- The members of the golf club have their following attributes and methods. We will create a class to represent any member of the club and consider appropriate constructors or setter methods to set the details and a method display() to display the details of the members.
Methodology:-
•	Class Date
               Class date contains the following methods getday(), getmonth(), getyear() which   	returns the day, month and year and a constructor which we will use to take input 	and a method displayDate() to display the date.
 

•	Class Member
               It contains the methods setDetails() asks the user for input and the details of members.
              The seniorhandicap() method shows the record of the senior members whose 	handicap score is less than 12.

And the method TeamBfemales() shows the record of the female senior members who are a part of TeamB.
Example : 
  
   And in the end a display() method in here displays the overall attributes of the members.
And a main() method from which we will implement all the above methods we made to do their respective operations and also contains the logic to printing info of joining before 7-April.


Conclusion:- In this way this methodology provides us with the way to store the details of the members of the golf club using the above classes and methods.